+++ src/cqterrain/book/__init__.py
@@ -1,4 +1,5 @@
 from .book import book
 from .book_random import book_random
 from .books import books
-from .bookcase import book_case as bookcase
+from .book_case import book_case as bookcase
+from .Bookcase import Bookcase